Add `experimental_readConfig` API
