+++ packages/wrangler/src/ai/createFinetune.tsx
@@ -2,7 +2,7 @@ import fs from "fs";
 import path from "path";
 import { FormData } from "undici";
 import { fetchResult } from "../cfetch";
-import { withConfig } from "../config";
+import { withConfig } from "../config/index";
 import { logger } from "../logger";
 import { requireAuth } from "../user";
 import { getErrorMessage } from "./utils";
