+++ packages/wrangler/src/d1/delete.ts
@@ -1,8 +1,8 @@
-import { printWranglerBanner } from "..";
 import { fetchResult } from "../cfetch";
 import { withConfig } from "../config";
 import { confirm } from "../dialogs";
 import { logger } from "../logger";
+import { printWranglerBanner } from "../update-check";
 import { requireAuth } from "../user";
 import { Name } from "./options";
 import { getDatabaseByNameOrBinding } from "./utils";
