+++ packages/wrangler/src/d1/execute.tsx
@@ -8,7 +8,6 @@ import Table from "ink-table";
 import md5File from "md5-file";
 import { Miniflare } from "miniflare";
 import { fetch } from "undici";
-import { printWranglerBanner } from "../";
 import { fetchResult } from "../cfetch";
 import { readConfig } from "../config";
 import { getLocalPersistencePath } from "../dev/get-local-persistence-path";
@@ -17,6 +16,7 @@ import { createFatalError, JsonFriendlyFatalError, UserError } from "../errors";
 import { logger } from "../logger";
 import { APIError, readFileSync } from "../parse";
 import { readableRelative } from "../paths";
+import { printWranglerBanner } from "../update-check";
 import { requireAuth } from "../user";
 import { renderToString } from "../utils/render";
 import * as options from "./options";
