+++ packages/wrangler/src/d1/export.ts
@@ -4,7 +4,6 @@ import { spinner, spinnerWhile } from "@cloudflare/cli/interactive";
 import chalk from "chalk";
 import { Miniflare } from "miniflare";
 import { fetch } from "undici";
-import { printWranglerBanner } from "..";
 import { fetchResult } from "../cfetch";
 import { readConfig } from "../config";
 import { getLocalPersistencePath } from "../dev/get-local-persistence-path";
@@ -12,6 +11,7 @@ import { UserError } from "../errors";
 import { logger } from "../logger";
 import { APIError } from "../parse";
 import { readableRelative } from "../paths";
+import { printWranglerBanner } from "../update-check";
 import { requireAuth } from "../user";
 import { Name } from "./options";
 import { getDatabaseByNameOrBinding, getDatabaseInfoFromConfig } from "./utils";
