+++ packages/wrangler/src/deployments.ts
@@ -9,8 +9,9 @@ import { UserError } from "./errors";
 import { mapBindings } from "./init";
 import { logger } from "./logger";
 import * as metrics from "./metrics";
+import { printWranglerBanner } from "./update-check";
 import { requireAuth } from "./user";
-import { getScriptName, printWranglerBanner } from ".";
+import { getScriptName } from "./utils/getScriptName";
 import type { Config } from "./config";
 import type { WorkerMetadataBinding } from "./deployment-bundle/create-worker-upload-form";
 import type { ServiceMetadataRes } from "./init";
