+++ packages/wrangler/src/generate/index.ts
@@ -2,12 +2,12 @@ import fs from "node:fs";
 import path from "node:path";
 import { execa } from "execa";
 import { getC3CommandFromEnv } from "../environment-variables/misc-variables";
-import { UserError } from "../errors";
+import { CommandLineArgsError, UserError } from "../errors";
 import { cloneIntoDirectory, initializeGit } from "../git-client";
-import { CommandLineArgsError, printWranglerBanner } from "../index";
 import { initHandler } from "../init";
 import { logger } from "../logger";
 import { getPackageManager } from "../package-manager";
+import { printWranglerBanner } from "../update-check";
 import * as shellquote from "../utils/shell-quote";
 import type {
 	CommonYargsArgv,
