+++ packages/wrangler/src/pages/secret/index.ts
@@ -6,11 +6,11 @@ import { findWranglerToml, readConfig } from "../../config";
 import { getConfigCache } from "../../config-cache";
 import { confirm, prompt } from "../../dialogs";
 import { FatalError } from "../../errors";
-import { printWranglerBanner } from "../../index";
 import isInteractive from "../../is-interactive";
 import { logger } from "../../logger";
 import * as metrics from "../../metrics";
 import { parseJSON, readFileSync } from "../../parse";
+import { printWranglerBanner } from "../../update-check";
 import { requireAuth } from "../../user";
 import { readFromStdin, trimTrailingWhitespace } from "../../utils/std";
 import { PAGES_CONFIG_CACHE_FILENAME } from "../constants";
