+++ packages/wrangler/src/triggers/index.ts
@@ -1,8 +1,10 @@
 import { processAssetsArg } from "../assets";
 import { readConfig } from "../config";
-import { getScriptName, isLegacyEnv, printWranglerBanner } from "../index";
 import * as metrics from "../metrics";
+import { printWranglerBanner } from "../update-check";
 import { requireAuth } from "../user";
+import { getScriptName } from "../utils/getScriptName";
+import { isLegacyEnv } from "../utils/isLegacyEnv";
 import triggersDeploy from "./deploy";
 import type {
 	CommonYargsArgv,
