+++ packages/wrangler/src/type-generation/index.ts
@@ -6,8 +6,7 @@ import { findWranglerToml, readConfig } from "../config";
 import { getEntry } from "../deployment-bundle/entry";
 import { getNodeCompatMode } from "../deployment-bundle/node-compat";
 import { getVarsForDev } from "../dev/dev-vars";
-import { UserError } from "../errors";
-import { CommandLineArgsError } from "../index";
+import { CommandLineArgsError, UserError } from "../errors";
 import { logger } from "../logger";
 import { parseJSONC } from "../parse";
 import { printWranglerBanner } from "../update-check";
