+++ packages/wrangler/src/utils/getScriptName.ts
@@ -1,0 +1,20 @@
+import { CommandLineArgsError } from "../errors";
+import { isLegacyEnv } from "./isLegacyEnv";
+import type { Config } from "../config";
+
+export function getScriptName(
+	args: { name: string | undefined; env: string | undefined },
+	config: Config
+): string | undefined {
+	if (args.name && isLegacyEnv(config) && args.env) {
+		throw new CommandLineArgsError(
+			"In legacy environment mode you cannot use --name and --env together. If you want to specify a Worker name for a specific environment you can add the following to your wrangler.toml config:" +
+				`
+    [env.${args.env}]
+    name = "${args.name}"
+    `
+		);
+	}
+
+	return args.name ?? config.name;
+}
