+++ packages/wrangler/src/utils/isLegacyEnv.ts
@@ -1,0 +1,7 @@
+import type { Config } from "../config";
+
+export function isLegacyEnv(config: Config): boolean {
+	// We only read from config here, because we've already accounted for
+	// args["legacy-env"] in https://github.com/cloudflare/workers-sdk/blob/b24aeb5722370c2e04bce97a84a1fa1e55725d79/packages/wrangler/src/config/validation.ts#L94-L98
+	return config.legacy_env;
+}
